Turn bonanza_builder into a worker
Instead of having a gRPC server integrated into bonanza_builder and
letting clients connect to it directly, this change alters it so that it
behaves like a worker that connects to the scheduler. bonanza_bazel will
now submit builds to the scheduler. This makes it easy to spin up pools
of bonanza_builder systems.