Add Message.OutgoingReference()
I was initially hesitant towards adding a function like this, because
there are also some places where we only care about the index of the
reference. Not the reference itself. It turns out that this is only
needed in some rare circumstances. In most of the high level code we
don't, so it's worth adding a shorthand for that.