Add Dereferencer
If we want to be able to extract child objects out of a
Message without necessarily calling into storage, we can't
just be calling into ParsedObjectReader directly. Instead, we
should have something that takes the OutgoingReferences of a
Message into consideration directly. To achieve this, we add a
new Dereferencer interface.

There are still many places in the tree where we should be
using Dereferencer instead of ParsedObjectReader.