Reimplement reposFilePropertiesResolver to not use DirectoryCluster
This means it will fetch objects more selectively. Instead of fetching
all Leaves messages at once, it will only fetch them for the final
directory in which the file to be looked up is contained.

This also means that we get closer to exclusively using Dereferencer as
part of the analyzer.