Pull TReference upwards
This change alters pkg/model/analysis, so that the implementation of
Computer has TReference parameterized. This will make it possible to
introduce a custom reference type that is capable of holding on to the
contents of objects that haven't been written to storage just yet. This
will allow us to run writes to storage asynchronously from the actual
evaluation.

While there, rework pkg/model/parser so that we can actually instantiate
caches as part of bonanza_worker and bonanza_builder. This makes builds
run faster due to a reduction in I/O against storage.