Completely pull TReference and TMetadata out of BaseComputer
Right now we execute all functions sequentially and synchronously.
Eventually we want to make it so that writing data to storage can simply
happen asynchronously. It should be possible for functions to execute in
a way that they can immediately use data yielded by previous functions,
without downloading the data from storage again.

To solve this, we should add "fatter" references that either include a
copy of the data or load it from storage on demand.