Eliminate the object_pb.Contents message
I initially added it, because I wanted to keep the option open of adding
more fields to it, but this turns out to be pointless. The hash in an
object's reference spans the entire contents of the object, meaning that
there needs to be some deterministic binary representation of objects.
Let's therefore use a simple byte slice.